// PortalAutoservicio/src/app/models/resultados-edd-preg.ts
export class SeccionesEddPreg {
    TOTALPUNTOSOBTENIDOS: string;
    CALIFICACIONPARCIAL: string;
    RESULTADOSPREGUNTASEDDDTO: ResultadosPreguntas[];
}


export class ResultadosPreguntas{
    IDPREGUNTA: string;
    DESCRIPCIONPREGUNTA: string;
    IDRESPUESTA: string;
    DSRESPUESTA: string;
}